+++ test/download_utils_test.dart
@@ -80,4 +80,47 @@ void main() {
           'https://github.com/daniel-maxhari/dynamic_text_highlighting/blob/master/subdir/LICENSE');
     });
   });
+
+  group('UrlChecker', () {
+    test('problem: invalid', () async {
+      final status = await UrlChecker().checkStatus('htp://pub.dev/');
+      expect(status.isInvalid, true);
+      expect(status.isInternal, false);
+      expect(status.isSecure, false);
+      expect(status.exists, false);
+      expect(status.getProblemCode(packageIsKnownInternal: false), 'invalid');
+      expect(status.getProblemCode(packageIsKnownInternal: true), 'invalid');
+    });
+
+    test('problem: internal', () async {
+      final status = await UrlChecker().checkStatus('https://pub.dev/');
+      expect(status.isInvalid, false);
+      expect(status.isInternal, true);
+      expect(status.isSecure, true);
+      expect(status.exists, true);
+      expect(status.getProblemCode(packageIsKnownInternal: false), 'internal');
+      expect(status.getProblemCode(packageIsKnownInternal: true), isNull);
+    });
+
+    test('problem: insecure', () async {
+      final status = await UrlChecker().checkStatus('http://pub.dev/');
+      expect(status.isInvalid, false);
+      expect(status.isInternal, true);
+      expect(status.isSecure, false);
+      expect(status.exists, true);
+      expect(status.getProblemCode(packageIsKnownInternal: false), 'internal');
+      expect(status.getProblemCode(packageIsKnownInternal: true), 'insecure');
+    });
+
+    test('problem: missing', () async {
+      final status = await UrlChecker()
+          .checkStatus('https://github.com/dart-lang/pub-dev/missing-url');
+      expect(status.isInvalid, false);
+      expect(status.isInternal, false);
+      expect(status.isSecure, true);
+      expect(status.exists, false);
+      expect(status.getProblemCode(packageIsKnownInternal: false), 'missing');
+      expect(status.getProblemCode(packageIsKnownInternal: true), 'missing');
+    });
+  });
 }