+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.20.0
 
 * Fix SDK version parsing issue (also reading `stdout` for newer SDKs).
+* `Summary.urlProblems` lists the URLs that should be displayed differently on `pub.dev`.
 * **BREAKING CHANGES**
   * `UrlStatus` is converted to a class with fields.
   * `UrlChecker` internal cache is removed (incl. `maxCacheSize`, `existsInCache`, `markExistsInCache`).